Updated file locations/formatting
