+++ src/CoreAsset.php
